Fix hydration issues appearing in Sanity Studio caused by route group layouts
