+++ src/app/layout.tsx
@@ -1,15 +1,6 @@
-import Header from "@/components/header";
-import Banner from "@/components/banner";
-import CookiesBanner from "@/components/cookies";
-import GoogleAnalytics from "@/components/google-analytics";
-import { urlFor } from "./actions/sanityClient";
 import "../styles/app.scss";
 import type { Metadata } from "next";
-import { cookies } from "next/headers";
-import { getGlobalContent } from "./actions/sanityClient";
-import { Suspense } from "react";
-import { MSWInitializer } from "@/components/msw-initialiser";
-import Head from "next/head";
+import { getGlobalContent, urlFor } from "./actions/sanityClient";
 
 export const globalContent = await getGlobalContent();
 export async function generateMetadata(): Promise<Metadata> {
@@ -22,49 +13,18 @@ export async function generateMetadata(): Promise<Metadata> {
   };
 }
 
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const environment = process.env.NEXT_PUBLIC_ENVIRONMENT;
-  const cookieStore = cookies();
-  const isShowCookie = cookieStore.get("isShowCookie")?.value || true;
-  const isConsentCookie = cookieStore.get("isConsentCookie")?.value || false;
-
   if (typeof window !== "undefined") {
     const govUk = require("govuk-frontend");
     govUk.initAll();
   }
-
   return (
     <html lang="en">
-      <body>
-        {isShowCookie == true && <CookiesBanner />}
-        {isConsentCookie &&
-          environment !== "development" &&
-          globalContent?.googleAnalytics && (
-            <GoogleAnalytics gaId={globalContent?.googleAnalytics} />
-          )}
-        <a
-          href="#main"
-          className="govuk-skip-link"
-          data-module="govuk-skip-link"
-        >
-          Skip to main content
-        </a>
-        <Header globalConfig={globalContent} />
-        <Banner globalConfig={globalContent} />
-        <main className="layout-wrap" id="main">
-          {process.env.NEXT_PUBLIC_API_MOCKING === "enabled" ? (
-            <MSWInitializer>
-              <Suspense>{children}</Suspense>
-            </MSWInitializer>
-          ) : (
-            <Suspense>{children}</Suspense>
-          )}
-        </main>
-      </body>
+      <body>{children}</body>
     </html>
   );
 }
